Add: Styles section to the browse mode sidebar.
